[PK-227] Implementation of dynamic value setting to cursor Making adjustments caused by table classes re-architecture
